requirements.txt
Django==4.2.1
djangorestframework==3.14.0
drf-yasg==1.21.5
pytest==7.3.1
